+++ packages/apps/kadena-docs/src/components/Layout/Full/Full.tsx
@@ -1,6 +1,6 @@
 import { Heading } from '@kadena/react-components';
 
-import { Article, Aside, CodeBackground, Content } from '../Code/styles';
+import { Article, Aside, CodeBackground, Content } from '../components';
 
 import { AsideLink } from './AsideLink';
 import { AsideList, StickyAside, StickyAsideWrapper } from './style';
