+++ packages/apps/kadena-docs/src/components/Layout/components/index.ts
@@ -3,4 +3,5 @@ export * from './Footer';
 export * from './DocsLogo';
 export * from './Main';
 export * from './styles';
+export * from './AsideStyles';
 export * from './TitleHeader';
